added select and input onchange
